chatbot_backend/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
langchain-core
sentence-transformers
torch
faiss-cpu
pypdf
pandas
geopandas
python-docx
python-dotenv
python-multipart
markdown
tiktoken
transformers>=4.41.1
huggingface_hub>=0.23.0
scikit-learn
numpy
datasets
transformers[torch]
